frontend/Agro2Watch/backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
aiofiles==23.2.1

# Authentication & Security
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP Client
httpx==0.25.2

# Environment & Configuration
python-dotenv==1.0.0

# Firebase Admin SDK (required at import time even in mock mode)
firebase-admin==6.2.0

# Optional: Production dependencies (uncomment when needed)
# google-cloud-firestore==2.13.1
# tensorflow==2.15.0
# pillow==10.1.0
# numpy==1.24.3
# opencv-python==4.8.1.78

# Optional: Alternate Flask-based server (only if using server.py)
Flask==2.3.3
Flask-Cors==4.0.0

# Lightweight image/Numpy stack used by utils.image_processing
Pillow==10.1.0
numpy==1.24.3

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1